Add integration tests for teacher forcing fixes
- test_full_training_100_steps: Verifies 100 training steps complete without OOM or type errors
- test_checkpoint_save_and_load: Verifies model weights can be saved and restored
- Both tests use eager execution to avoid graph mode issues
- All 8 tests now passing (3 type compatibility, 3 gradient flow, 2 integration)

Also fixed type mismatch by casting literal 0 to int64 in update_tf_ratio()

Closes task-1773015095-768c

+++ src/music_generation/train.py
@@ -123,7 +123,7 @@ class MelGeneratorTraining(tf.keras.Model):
         """Update teacher forcing ratio based on current training step."""
         # Apply warmup: keep initial ratio until warmup_steps
         # Cast warmup_steps to int64 for type compatibility with training_step
-        step_after_warmup = tf.maximum(0, self.training_step - tf.cast(self.warmup_steps, tf.int64))
+        step_after_warmup = tf.maximum(tf.cast(0, tf.int64), self.training_step - tf.cast(self.warmup_steps, tf.int64))
         
         # Compute new ratio using exponential schedule
         new_ratio = exponential_tf_schedule(
